+++ client/data/constants.js
@@ -5,7 +5,7 @@
 // classes
 
 // ... other things?
-export const INPUT_ID = 'name-input'
-export const USER_ACTION='add-product'
-export const OUTPUT_ID='mySelect'
-export const BUTTONCLICK_EVENT='click'
+export const INPUT_ID = 'name-input';
+export const USER_ACTION = 'add-product';
+export const OUTPUT_ID = 'mySelect';
+export const BUTTONCLICK_EVENT = 'click';
