+++ client/src/components/addcomponent.js
@@ -1,0 +1,5 @@
+export const renderOption = (value) => {
+  const newOption = document.createElement('option');
+  newOption.innerText = value;
+  return newOption;
+};
