--- client/src/components/component.js
@@ -1,5 +1,0 @@
-export const renderOption = (value) => {
-    const newOption = document.createElement('option');
-    newOption.innerText = value;
-    return newOption;
-}
