+++ client/src/handlers/addhandler.js
@@ -1,0 +1,8 @@
+import { INPUT_ID, OUTPUT_ID } from '../../data/constants.js';
+import { renderOption } from '../components/addcomponent.js';
+
+export const addanItem = () => {
+  const newItemValue = document.getElementById(INPUT_ID).value;
+  const newOption = renderOption(newItemValue);
+  document.getElementById(OUTPUT_ID).prepend(newOption);
+};
