+++ client/src/init.js
@@ -7,7 +7,7 @@
   x.add(option, x[0]);
 });
 */
-import { USER_ACTION } from "../data/constants.js";
-import { click}from "./listeners/addlistener.js"
+import { USER_ACTION } from '../data/constants.js';
+import { click } from './listeners/addlistener.js';
 
 click(document.getElementById(USER_ACTION));
