+++ client/src/listeners/addlistener.js
@@ -1,6 +1,6 @@
-import { BUTTONCLICK_EVENT } from "../../data/constants.js";
-import {addanItem}from "../handlers/handler.js";
+import { BUTTONCLICK_EVENT } from '../../data/constants.js';
+import { addanItem } from '../handlers/addhandler.js';
 
-export const click=( value)=>{
-    value.addEventListener(BUTTONCLICK_EVENT,addanItem);
-}
+export const click = (value) => {
+  value.addEventListener(BUTTONCLICK_EVENT, addanItem);
+};
